# Docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Next.js API (Feature Requests, Profiles, Personas, Contentstack)",
    "description": "API for authentication (Supabase), feature requests (with comments/votes and file uploads), user profiles & personas, Contentstack publishing helpers, and internal log ingestion.",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/update-demo-web-in-cs": {
      "put": {
        "tags": [
          "Contentstack"
        ],
        "summary": "Update & publish custom demos in Contentstack",
        "description": "Updates the `demos` field on a Contentstack `custom_demos` entry and publishes it. Requires `UPLOAD_DEMO_WEBSITES` permission.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentstackUpdateRequestDemos"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Demos updated & published.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentstackUpdateResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (missing permission).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentstackUpdateResult"
                }
              }
            }
          },
          "default": {
            "description": "Upstream Contentstack error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentstackUpdateResult"
                }
              }
            }
          }
        },
        "operationId": "putApiUpdateDemoWebInCs"
      },
      "delete": {
        "tags": [
          "Contentstack"
        ],
        "summary": "Delete a demo (by sending new demos array) & publish",
        "description": "Same as update, but used by UI to 'delete' a demo website by sending the updated `demos` array. Requires `UPLOAD_DEMO_WEBSITES` permission.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentstackUpdateRequestDemos"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Demo deleted & entry published.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentstackUpdateResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (missing permission).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentstackUpdateResult"
                }
              }
            }
          },
          "default": {
            "description": "Upstream Contentstack error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentstackUpdateResult"
                }
              }
            }
          }
        },
        "operationId": "deleteApiUpdateDemoWebInCs"
      }
    },
    "/api/update-video-library-in-cs": {
      "put": {
        "tags": [
          "Contentstack"
        ],
        "summary": "Update & publish video library in Contentstack",
        "description": "Updates the `videos` field on a Contentstack `video_library` entry and publishes it. Requires `UPLOAD_VIDEO_LIBRARY` permission.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentstackUpdateRequestVideos"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Videos updated & published.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentstackUpdateResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (missing permission).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentstackUpdateResult"
                }
              }
            }
          },
          "default": {
            "description": "Upstream Contentstack error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentstackUpdateResult"
                }
              }
            }
          }
        },
        "operationId": "putApiUpdateVideoLibraryInCs"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sb-<project-ref>-auth-token",
        "description": "Supabase SSR session cookie (project-specific name like `sb-<project-ref>-auth-token`)."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Bearer access token (e.g., Supabase access_token)."
      },
      "internalLogSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-internal-log-secret",
        "description": "Shared secret used to authorize internal log ingestion."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {}
        },
        "required": [
          "error"
        ]
      },
...
      "ContentstackAssetRef": {
        "description": "Contentstack asset reference. In update payloads this is often just the asset UID string.",
        "oneOf": [
          {
            "type": "string",
            "description": "Asset UID"
          },
          {
            "type": "object",
            "properties": {
              "uid": {
                "type": "string"
              }
            },
            "required": [
              "uid"
            ],
            "additionalProperties": true
          }
        ],
        "nullable": true
      },
      "DemoWebsite": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "se_name": {
            "type": "string"
          },
          "date_posted": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "image": {
            "$ref": "#/components/schemas/ContentstackAssetRef"
          },
          "link": {
            "type": "object",
            "properties": {
              "title": {
                "type": "string"
              },
              "href": {
                "type": "string"
              }
            },
            "required": [
              "title",
              "href"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "title",
          "description",
          "se_name",
          "link"
        ],
        "additionalProperties": false
      },
...
      "VideoItem": {
        "type": "object",
        "properties": {
          "video_url": {
            "type": "string",
            "format": "uri"
          },
          "video_file": {
            "$ref": "#/components/schemas/ContentstackAssetRef"
          },
          "thumbnail": {
            "$ref": "#/components/schemas/ContentstackAssetRef"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "se_name": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "description",
          "se_name"
        ],
        "additionalProperties": false
      },
...
      "ContentstackUpdateRequestDemos": {
        "type": "object",
        "required": [
          "entryUid",
          "demos"
        ],
        "properties": {
          "entryUid": {
            "type": "string"
          },
          "demos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DemoWebsite"
            }
          }
        },
        "additionalProperties": false
      },
      "ContentstackUpdateRequestVideos": {
        "type": "object",
        "required": [
          "entryUid",
          "videos"
        ],
        "properties": {
          "entryUid": {
            "type": "string"
          },
          "videos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VideoItem"
            }
          }
        },
        "additionalProperties": false
      },
      "ContentstackUpdateResult": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer"
          },
          "entry": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "error": {
            "type": "string",
            "nullable": true
          },
          "details": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "status"
        ],
        "additionalProperties": false
      },
...
    }
  }
}